+++ tailwind.config.js
@@ -1,7 +1,7 @@
 module.exports = {
 	purge: {
 		enabled: true,
-		content: ["./src/**/*.php"],
+		content: ['./src/**/*.php'],
 	},
 	darkMode: false, // or 'media' or 'class'
 	theme: {